pyproject.toml
[project]
name = "docmancer"
version = "0.1.0"
description = "A Python CLI tool for automated documentation generation and insertion."
authors = [
    {name = "Brody Clark",email = "brodyclark97@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
click = "^8.1.3"
httpx = ">=0.28.1,<0.29.0"
pyyaml = ">=6.0.2,<7.0.0"
dataclasses-json = "^0.6.7"
llama_cpp_python = "^0.3.9"
prompt_toolkit = "^3.0.51"
Pygments = "^2.19.1"
rich = "^14.0.0"
tree-sitter = "^0.24.0"
tree-sitter-c-sharp = "^0.23.1"
tree-sitter-cpp = "^0.23.4"
tree-sitter-javascript = "^0.23.1"
tree-sitter-python = "^0.23.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
docmancer = "docmancer.cli:main" # 'command_name = package.module:function'